+++ truffle/src/com.oracle.truffle.host/src/com/oracle/truffle/host/HostMethodScope.java
@@ -65,6 +65,7 @@ final class HostMethodScope {
 
     private static final ScopedObject[] EMTPY_SCOPE_ARRAY = new ScopedObject[0];
     private static final Unsafe UNSAFE = getUnsafe();
+    private static final Message MESSAGE_AS_HOST_OBJECT = Message.resolveExact(InteropLibrary.class, "asHostObject", Object.class);
 
     private ScopedObject[] scope;
     private int nextDynamicIndex;
@@ -246,7 +247,7 @@ final class HostMethodScope {
                                 "Alternatively, use Value.pin() to prevent a scoped object from being released after the host call completed.");
             }
             Object returnValue = library.send(d, message, args);
-            if (message.getReturnType() == Object.class && !(d instanceof PinnedObject)) {
+            if (message.getReturnType() == Object.class && !(d instanceof PinnedObject) && message != MESSAGE_AS_HOST_OBJECT) {
                 /*
                  * Object return type indicates for an interop message that any interop value may be
                  * returned.